[ constraints ] refactor: split blockers and elims from Syntax.Internal into separate modules
